Design RatingWidget - Empty Star
