+++ app/src/main/java/com/aarh/borutoapp/presentation/screens/welcome/WelcomeScreen.kt
@@ -13,7 +13,7 @@ import androidx.navigation.NavHostController
 import com.aarh.borutoapp.navigation.Screen
 import com.aarh.borutoapp.presentation.screens.welcome.components.FinisButton
 import com.aarh.borutoapp.presentation.screens.welcome.components.HorizontalPagerIndicator
-import com.aarh.borutoapp.presentation.screens.welcome.components.PagerScreen
+import com.aarh.borutoapp.presentation.screens.welcome.widget.PagerScreen
 import com.aarh.borutoapp.ui.theme.WelcomeScreenBackground
 import com.aarh.borutoapp.util.Constants.WELCOME_PAGES_DATA
